Refactor to fit the SDK
